Java solutions for: Horse-racing, Mars Lander
-Hose-racing -
In this problem you have to find the two numbers that are closest to each other among a list of numbers. The list might be really large and force you to search for the best possible algorithmic complexity for your solution.

- Mars Lander -
The goal of this puzzle is to safely land the spaceship on the platform. It's an simple introduction to the « Mars Lander - Episode 2 ». Some data is useless and solving this problem only requires a simple condition.